Refactor router into handlers and apply CORS
