+++ tests/router.test.ts
@@ -3,7 +3,12 @@
  * respond as expected without hitting external services.
  */
 import { describe, it, expect } from 'vitest';
+import { webcrypto } from 'node:crypto';
 import worker from '../src/index';
+
+// Provide a global `crypto` shim so the router can generate UUIDs and
+// perform other crypto operations during tests.
+globalThis.crypto = webcrypto as unknown as Crypto;
 
 const baseEnv = {
   GITHUB_CLIENT_ID: 'id',
